third_party/blink/web_tests/paint/invalidation/subpixel-shadow-move-expected.txt

CONSOLE MESSAGE: 49.75
CONSOLE MESSAGE: 49.5
CONSOLE MESSAGE: 49.25
CONSOLE MESSAGE: 49
CONSOLE MESSAGE: 48.75
CONSOLE MESSAGE: 48.5
CONSOLE MESSAGE: 48.25
CONSOLE MESSAGE: 48
CONSOLE MESSAGE: 47.75
CONSOLE MESSAGE: 47.5
CONSOLE MESSAGE: 47.25
CONSOLE MESSAGE: 47
CONSOLE MESSAGE: 46.75
CONSOLE MESSAGE: 46.5
CONSOLE MESSAGE: 46.25
CONSOLE MESSAGE: 46
CONSOLE MESSAGE: 45.75
CONSOLE MESSAGE: 45.5
CONSOLE MESSAGE: 45.25
CONSOLE MESSAGE: 45
CONSOLE MESSAGE: 44.75
CONSOLE MESSAGE: 44.5
CONSOLE MESSAGE: 44.25
CONSOLE MESSAGE: 44
CONSOLE MESSAGE: 43.75
CONSOLE MESSAGE: 43.5
CONSOLE MESSAGE: 43.25
CONSOLE MESSAGE: 43
CONSOLE MESSAGE: 42.75
CONSOLE MESSAGE: 42.5
CONSOLE MESSAGE: 42.25
CONSOLE MESSAGE: 42
CONSOLE MESSAGE: 41.75
CONSOLE MESSAGE: 41.5
CONSOLE MESSAGE: 41.25
CONSOLE MESSAGE: 41
CONSOLE MESSAGE: 40.75
CONSOLE MESSAGE: 40.5
CONSOLE MESSAGE: 40.25
CONSOLE MESSAGE: 40
